number-classification-django-rest-api/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
requests==2.31.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
